+++ src/encoder.rs
@@ -1,11 +1,11 @@
 // encoder.rs
 //
-// Copyright (c) 2019-2020  Minnesota Department of Transportation
+// Copyright (c) 2019-2021  Minnesota Department of Transportation
 //
 //! Encoder for Mapbox Vector Tile (MVT) geometry.
 //!
 use crate::error::Error;
-use pointy::Transform64;
+use pointy::Transform;
 
 #[derive(Copy, Clone, Debug)]
 enum Command {
@@ -43,9 +43,9 @@ pub enum GeomType {
 /// # Example
 /// ```
 /// # use mvt::{Error, GeomEncoder, GeomType};
-/// # use pointy::Transform64;
+/// # use pointy::Transform;
 /// # fn main() -> Result<(), Error> {
-/// let geom_data = GeomEncoder::new(GeomType::Point, Transform64::default())
+/// let geom_data = GeomEncoder::new(GeomType::Point, Transform::default())
 ///     .point(0.0, 0.0)
 ///     .point(10.0, 0.0)
 ///     .encode()?;
@@ -53,7 +53,7 @@ pub enum GeomType {
 /// ```
 pub struct GeomEncoder {
     geom_tp: GeomType,
-    transform: Transform64,
+    transform: Transform<f64>,
     x: i32,
     y: i32,
     cmd_offset: usize,
@@ -68,9 +68,9 @@ pub struct GeomEncoder {
 /// # Example
 /// ```
 /// # use mvt::{Error, GeomEncoder, GeomType};
-/// # use pointy::Transform64;
+/// # use pointy::Transform;
 /// # fn main() -> Result<(), Error> {
-/// let geom_data = GeomEncoder::new(GeomType::Point, Transform64::default())
+/// let geom_data = GeomEncoder::new(GeomType::Point, Transform::default())
 ///     .point(0.0, 0.0)
 ///     .point(10.0, 0.0)
 ///     .encode()?;
@@ -106,7 +106,7 @@ impl GeomEncoder {
     ///
     /// * `geom_tp` Geometry type.
     /// * `transform` Transform to apply to geometry.
-    pub fn new(geom_tp: GeomType, transform: Transform64) -> Self {
+    pub fn new(geom_tp: GeomType, transform: Transform<f64>) -> Self {
         GeomEncoder {
             geom_tp,
             transform,
@@ -246,7 +246,7 @@ mod test {
     // Examples from MVT spec:
     #[test]
     fn test_point() {
-        let v = GeomEncoder::new(GeomType::Point, Transform64::default())
+        let v = GeomEncoder::new(GeomType::Point, Transform::default())
             .point(25.0, 17.0)
             .encode()
             .unwrap()
@@ -256,7 +256,7 @@ mod test {
 
     #[test]
     fn test_multipoint() {
-        let v = GeomEncoder::new(GeomType::Point, Transform64::default())
+        let v = GeomEncoder::new(GeomType::Point, Transform::default())
             .point(5.0, 7.0)
             .point(3.0, 2.0)
             .encode()
@@ -267,7 +267,7 @@ mod test {
 
     #[test]
     fn test_linestring() {
-        let v = GeomEncoder::new(GeomType::Linestring, Transform64::default())
+        let v = GeomEncoder::new(GeomType::Linestring, Transform::default())
             .point(2.0, 2.0)
             .point(2.0, 10.0)
             .point(10.0, 10.0)
@@ -279,7 +279,7 @@ mod test {
 
     #[test]
     fn test_multilinestring() {
-        let v = GeomEncoder::new(GeomType::Linestring, Transform64::default())
+        let v = GeomEncoder::new(GeomType::Linestring, Transform::default())
             .point(2.0, 2.0)
             .point(2.0, 10.0)
             .point(10.0, 10.0)
@@ -295,7 +295,7 @@ mod test {
 
     #[test]
     fn test_polygon() {
-        let v = GeomEncoder::new(GeomType::Polygon, Transform64::default())
+        let v = GeomEncoder::new(GeomType::Polygon, Transform::default())
             .point(3.0, 6.0)
             .point(8.0, 12.0)
             .point(20.0, 34.0)
@@ -307,7 +307,7 @@ mod test {
 
     #[test]
     fn test_multipolygon() {
-        let v = GeomEncoder::new(GeomType::Polygon, Transform64::default())
+        let v = GeomEncoder::new(GeomType::Polygon, Transform::default())
             // positive area => exterior ring
             .point(0.0, 0.0)
             .point(10.0, 0.0)