+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [Unreleased]
 
+### Changed
+* Moved `BBox` to `pointy` crate
+
 ## [0.7.0] - 2020-09-29
 ### Changed
 * Replaced `geom` module with `pointy` crate dependency